fix: honeycomb-beeline bug workaround
Honeycomb monkeypatches `pg.Client.prototype.query` and wraps the last argument
in a function. This breaks the pg-query-stream integration. An open issue is
linked at the end of this message.

In the meantime, we can pass a dummy argument to `query()`, which will be
wrapped instead.

Issue: https://github.com/honeycombio/beeline-nodejs/issues/173